ai-video-bot/requirements.txt
openai
requests
moviepy
gTTS
google-api-python-client
python-dotenv
pytrends
feedparser